#12 Update custom post feature, rollback customModelPost to CustomPost, add WpOrmException
