--- src/Scopes/CustomPostAddTypeScope.php
@@ -1,29 +1,0 @@
-<?php
-/**
- * Copyright (c) 2024 Dimitri BOUTEILLE (https://github.com/dimitriBouteille)
- * See LICENSE.txt for license details.
- *
- * Author: Dimitri BOUTEILLE <bonjour@dimitri-bouteille.fr>
- */
-
-namespace Dbout\WpOrm\Scopes;
-
-use Dbout\WpOrm\Builders\PostBuilder;
-use Dbout\WpOrm\Models\CustomPost;
-use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Scope;
-
-class CustomPostAddTypeScope implements Scope
-{
-    /**
-     * @param PostBuilder $builder
-     * @inheritDoc
-     */
-    public function apply(Builder $builder, Model $model)
-    {
-        /** @var CustomPost $model */
-        $type = $model->getPostType();
-        $builder->whereTypes($type);
-    }
-}
